\name{EPA}
\alias{EPA}
\title{
...
}
\description{

}
\usage{
EPA(dat, alpha.reject = 0.05, write.result = T)
}

\arguments{
  \item{dat}{}
  \item{alpha.reject}{}
  \item{write.result}{}
}
\details{

}
\value{

}
\references{

}
\author{
  Bill Shipley
}
\note{

}



\seealso{

}
\examples{

}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
