\name{MCX2}
\alias{MCX2}

\title{
...
}
\description{

}
\usage{
MCX2(model.df, n.obs, model.chi.square, n.sim = 10000) 
}

\arguments{
  \item{x}{}
  \item{model.df}{}
  \item{n.obs}{}
  \item{model.chi.square}{}
  \item{n.sim}{}
}
\details{

}
\value{

}
\references{

}
\author{
  Bill Shipley
}
\note{

}



\seealso{

}
\examples{

}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
