\name{orient.phaseII.1}
\alias{orient.phaseII.1}
\title{
title
}
\description{
description
}
\usage{
orient.phaseII.1(cgraph, nvars)
}
\arguments{
  \item{cgraph}{
need description
}
  \item{nvars}{
need description
}
}
\details{
details
}
\value{
value
}
\references{
ref
}
\author{
Bill Shipley
}
\seealso{
see also
}
\examples{
# None
}

