\name{orient.phasell.1}
\alias{orient.phasell.1}

\title{
...
}
\description{

}
\usage{
orient.phasell.1(cgraph, nvars) 
}

\arguments{
  \item{cgraph}{}
  \item{nvars}{}
}
\details{

}
\value{

}
\references{

}
\author{
  Bill Shipley
}
\note{

}



\seealso{

}
\examples{

}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
