\name{Pcor.prob}
\alias{Pcor.prob}

\title{
...
}
\description{

}
\usage{
Pcor.prob(dat, x, y, Q)
}

\arguments{
  \item{dat}{}
  \item{x}{}
  \item{y}{}
  \item{Q}{}
}
\details{

}
\value{

}
\references{

}
\author{
  Bill Shipley
}
\note{

}



\seealso{

}
\examples{

}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
