\name{shipley.test2}
\alias{shipley.test2}

\title{
...
}
\description{

}
\usage{
shipley.test2(amat, S, n)
}

\arguments{
  \item{amat}{}
  \item{S}{}
  \item{n}{}
}
\details{

}
\value{

}
\references{

}
\author{
  Bill Shipley
}
\note{

}



\seealso{

}
\examples{

}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
