\name{Causal.Inference}
\alias{Causal.Inference}

\title{
No title yet
}
\description{
No description yet
}
\usage{
Causal.Inference(dat, alpha.reject = 0.05, write.result = T)
}

\arguments{
  \item{dat}{...}
  \item{alpha.reject}{...}
  \item{write.result}{...}
}
\details{

}
\value{

}
\references{

}
\author{
  Bill Shipley
}
\note{

}



\seealso{

}
\examples{

}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
