\name{dsep.test}
\alias{dsep.test}

\title{
...
}
\description{

}
\usage{
dsep.test(amat, S, n, only.null = F)
}

\arguments{
  \item{amat}{}
  \item{s}{}
  \item{n}{}
  \item{only.null}{}
}
\details{

}
\value{

}
\references{

}
\author{
  Bill Shipley
}
\note{

}



\seealso{

}
\examples{

}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
