\name{EPA}
\alias{EPA}
\title{
Title
}
\description{
Description
}
\usage{
EPA(dat, alpha.reject = 0.05, write.result = T)
}
\arguments{
  \item{dat}{
need description
}
  \item{alpha.reject}{
need description
}
  \item{write.result}{
need description
}
}
\details{
need details
}
\value{
need value
}
\references{
need ref
}
\author{
Bill Shipley
}
\seealso{
need see also
}
\examples{
# need examples
}

