\name{Exploratory.path.analysis}
\alias{Exploratory.path.analysis}

\title{
...
}
\description{

}
\usage{
Exploratory.path.analysis(dat, upper.bound = 0.5, significance.level = 0.05) 
}

\arguments{
  \item{dat}{}
  \item{upper.bound}{}
  \item{significance.level}{}
}
\details{

}
\value{

}
\references{

}
\author{
  Bill Shipley
}
\note{

}



\seealso{

}
\examples{

}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
