\name{Observed.Equivalent.DAG}
\alias{Observed.Equivalent.DAG}

\title{
...
}
\description{

}
\usage{
Observed.Equivalent.DAG(full.DAG, latents = NA)
}

\arguments{
  \item{full.DAG}{}
  \item{latents}{}
}
\details{

}
\value{

}
\references{

}
\author{
  Bill Shipley
}
\note{

}



\seealso{

}
\examples{

}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
