\name{Observed.Equivalent.DAG}
\alias{Observed.Equivalent.DAG}

\title{
Observed.Equivalent.DAG
}
\description{
This function takes, as input, a Directed Acyclic Graph (DAG) as well as some variables (nodes) that you cannot directly measure (latent variables).  It then outputs a new partially oriented graph, containing only those variables in the original DAG that are observed, that is equivalent to the original DAG in terms of its d-separation claims.  If this partially oriented graph can be completely oriented to make a new DAG, then this too is output.  Any data generated by the original DAG must also agree with this new DAG in terms of the resulting pattern of (partial) dependencies.  In particular, if the data were generated by the original DAG then these data will also agree with the covariance structure of the new DAG.
}
\usage{
Observed.Equivalent.DAG(full.DAG, latents = NA)
}
\arguments{
  \item{full.DAG}{a binary matrix describing the DAG; usually from the DAG function}
  \item{latents}{a vector containing the names of the variables in the full.DAG that are latent (unobserved).  These names must agree with thos in full.DAG}
}
\details{
The input DAG is a binary matrix that is usually created via the DAG function in the ggm library.  
}
\author{
  Bill Shipley
}
\seealso{
\code{\link{DAG}}
}
\examples{
#test.dag<-DAG(x2~x1,x3~x2,x4~x2,x5~x3+x4)
#Observed.Equivalent.DAG(test.dag,"x2")
#the original DAG is: 
#x1 -> x2 
#x2 -> x3 
#x2 -> x4 
#x3 -> x5 
#x4 -> x5 
#latent variable(s):  x2 
#_____________________ 
#Equivalent partially oriented graph involving only the observed variables: 
#x1 -- x3 
#x1 -- x4 
#x3 -- x4 
#x3 -> x5 
#x4 -> x5 
#One *possible* equivalent DAG involving only the observed variables: 
#x1 -> x3 
#x1 -> x4 
#x3 -> x4 
#x3 -> x5 
#x4 -> x5 

}
