\name{orient.graph}
\alias{orient.graph}

\title{
...
}
\description{

}
\usage{
orient.graph(cgraph, nvars)
}

\arguments{
  \item{cgraph}{}
  \item{nvars}{}
}
\details{

}
\value{

}
\references{

}
\author{
  Bill Shipley
}
\note{

}



\seealso{

}
\examples{

}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
