\name{orient.phaseII.1}
\alias{orient.phaseII.1}
\title{
Here is a title
}
\description{
Need to clarify what it does
}
\usage{
orient.phaseII.1(cgraph, nvars) 
}

\arguments{
  \item{cgraph}{}
  \item{nvars}{}
}
\details{

}
\value{

}
\references{

}
\author{
  Bill Shipley
}
\note{

}



\seealso{

}
\examples{

}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
