\name{vanishing.tetrads}
\alias{vanishing.tetrads}

\title{
...
}
\description{

}
\usage{
vanishing.tetrads(dat, sig = 0.05)
}

\arguments{
  \item{dat}{}
  \item{sig}{}
}
\details{

}
\value{

}
\references{

}
\author{
  Bill Shipley
}
\note{

}



\seealso{

}
\examples{

}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
